SwarmNet - PyTorch/requirements.txt
torch~=1.11.0
PyYAML~=6.0
pandas~=1.4.2
numpy~=1.22.3
matplotlib~=3.5.1